refactor: move types under error_store
