+++ plugins/browser/ts/error_store.ts
@@ -1,8 +1,6 @@
-import { createStore, push, withAdder, type Store, type WithAdder } from '@just-web/states'
+import { createStore, push, withAdder } from '@just-web/states'
 import { pick } from 'type-plus'
-import type { ReadonlyErrorStore } from './types.js'
-
-export interface ErrorStore extends Store<Error[]>, WithAdder<Error> {}
+import type { ErrorStore, ReadonlyErrorStore } from './error_store.types.js'
 
 export function createErrorStore(): ErrorStore {
 	return withAdder(createStore<Error[]>([]), push)
