+++ plugins/browser/ts/error_store.types.ts
@@ -1,0 +1,6 @@
+import type { ReadonlyStore } from '@just-web/states'
+import { type Store, type WithAdder } from '@just-web/states'
+
+export interface ErrorStore extends Store<Error[]>, WithAdder<Error> {}
+
+export interface ReadonlyErrorStore extends ReadonlyStore<Error[]> {}
