+++ plugins/browser/ts/index.ts
@@ -1,4 +1,4 @@
-export type { ReadonlyErrorStore } from './types.js'
 export * from './browserPlugin.js'
 export { default } from './browserPlugin.js'
+export type { ErrorStore, ReadonlyErrorStore } from './error_store.types.js'
 export * from './errors.js'
