--- plugins/browser/ts/types.ts
@@ -1,3 +1,0 @@
-import type { ReadonlyStore } from '@just-web/states'
-
-export interface ReadonlyErrorStore extends ReadonlyStore<Error[]> {}
